**Hello reviewer**,

In this pull request we followed the next steps:

## Alter table owners ##
- Altered table owners to add new column email.

## Add data to perform analyze on visits table ##
- We used 3 times the query INSERT INTO visits (animals_id, vets_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp.  We used this querie three times to get the performance the activity asked for.

## Add data to perform analyze on owners table ##
- We used 3 times the query  INSERT INTO owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com'; We used two times the query to get the performance the activity asked for.

## Executed queries to get actual performance ##

1. explain analyze SELECT COUNT(*) FROM visits where animal_id = 4;

<img width="856" alt="1seg query" src="https://user-images.githubusercontent.com/29135748/229620508-4d3fe7cf-f118-4a16-a5bb-9b27af379bb9.png">

2. explain analyze SELECT * FROM visits where vet_id = 2;

<img width="782" alt="2nd query" src="https://user-images.githubusercontent.com/29135748/229620879-3014eba2-016c-4fdf-bc07-6f5640d118a6.png">

3. SELECT * FROM owners where email = 'owner_18327@mail.com';

<img width="750" alt="owners 1seg query" src="https://user-images.githubusercontent.com/29135748/229620857-b6c4b102-2de1-4688-8820-493fc8aa59be.png">

## Create indexes to enhance performance on queries ##

1. CREATE INDEX animals_id_desc_index ON visits(animals_id DESC);

<img width="990" alt="Exect time with index animals_id" src="https://user-images.githubusercontent.com/29135748/229621704-8a30bc9d-e969-4c18-8ff4-4f82c9ee63e2.png">

This index **reduce the execution time from 1460ms to 60ms.**

2. CREATE INDEX vets_id_index ON visits(vets_id DESC);

<img width="869" alt="120ms reduction" src="https://user-images.githubusercontent.com/29135748/229621764-0bd42e19-d9d6-4f7c-b65b-e224c8e3e292.png">

This query performance was just a **little improve from 964ms to 800ms.**

3. CREATE INDEX owners_email_index ON owners(email DESC);

<img width="771" alt="owner_mail_index" src="https://user-images.githubusercontent.com/29135748/229621751-03d0b679-23c2-4bdc-a9bd-303c36bd8613.png">

And the winner by far was this index, it was a **reduction from 1072 to 0,19 ms.**

**Thanks for reviewing our code, greeting @NelsonAraujo93  and @AlucardSanin ** 😄.